Compiles, with warnings
